requirements.txt
streamlit==1.32.0
python-dotenv==1.0.1
PyMuPDF==1.23.26
google-generativeai==0.3.2 